misc refactor

+++ src/Castlevania/Animation/AnimationFrame.h
@@ -6,10 +6,10 @@ class AnimationFrame
 {
 public:
 	AnimationFrame(Sprite *sprite, int time);
-	DWORD GetTime() { return time; }
+	float GetTime() { return time; }
 	Sprite *GetSprite() { return sprite; }
 
 private:
 	Sprite *sprite;
-	DWORD time;
+	float time;
 };
